Crate ibc_integration_test

IBC Relayer Integration Test Suite

Overview

The ibc-relayer-test crate provides the infrastructure and framework for writing end-to-end (E2E) tests that include the spawning of the relayer together with Cosmos full nodes running as child processes inside the tests.

Build Documentation

This documentation is best viewed as Rustdoc HTML pages. You can run the following command to build and view the documentation using cargo doc:

cargo doc -p ibc-integration-test --open

Installation

Other than Rust, the test suite assumes the gaiad binary is present in $PATH. You can install Gaia by either building from source, or load it using Cosmos.nix:

nix shell github:informalsystems/cosmos.nix#gaia5

Quick Start

See tests::example for an example test with access to a full test setup.

Modules

bootstrap

Helper functions for setting up test cases in an imperative way.

chain

Constructs for spawning and managing full nodes, e.g. the Gaia chains.

error

Error type used for the tests.

framework

Framework code for making it easier to write test cases.

ibc

Code that may belong to the ibc module, but are currently in this crate for easier review or maintenance.

prelude

Re-export of common constructs that are used by test cases.

relayer

Code that may belong to the ibc_relayer module, but are currently in this crate for easier review or maintenance.

tests

All test cases are placed within this module.

types

This module contains definitions of core data structures that are used in the test suite.

util

Utility and helper functions used in the tests.