Module ibc_integration_test::bootstrap

Helper functions for setting up test cases in an imperative way.

Normal test authors should have no need to call functions provided by the bootstrap module, as they are implicitly called by the framework constructs.

Advanced test authors with needs for more flexibility can call functions in the bootstrap module directly, so that they have more control of when exactly new chains and relayers should be spawned.

Modules

binary

Helper functions for bootstrapping constructs that involve two chains.

init

Functions for initializing each test at the beginning of a Rust test session.

nary

Experimental work to bootstrap N-ary chains for testing.

single

Helper functions for bootstrapping a single full node.