Module ibc_integration_test::bootstrap::binary::chain

Helper functions for bootstrapping two relayer chain handles with connected foreign clients.

Functions

add_chain_config

Generate ChainConfig from a running FullNode and add it to the relayer’s Config.

add_key_to_chain_handle

Add a wallet key to a ChainHandle’s key store.

add_keys_to_chain_handle

Add multiple wallets provided in TestWallets into the ChainHandle’s key store.

boostrap_chain_pair_with_nodes

Bootstraps two relayer chain handles with connected foreign clients.

boostrap_self_connected_chain

Work similary to boostrap_chain_pair_with_nodes, but bootstraps a single chain to be connected with itself.

bootstrap_foreign_client

Bootstrap a foreign client from ChainA to ChainB, i.e. the foreign client collects information from ChainA and submits them as transactions to ChainB.

new_registry

Create a new SharedRegistry that uses ProdChainHandle as the ChainHandle implementation.

pad_client_ids

save_relayer_config

Save a relayer’s Config to the filesystem to make it accessible through external CLI.

spawn_chain_handle

Spawn a new chain handle using the given SharedRegistry and FullNode.