Module ibc_integration_test::bootstrap::binary::channel

Helper functions for bootstrapping a channel between two chains.

Functions

bootstrap_channel

Create a new ConnectedChannel between two chains using foreign clients with initialized client IDs.

bootstrap_channel_with_chains

Create a new ConnectedChannel based on the provided ConnectedChains.

bootstrap_channel_with_connection

Create a new ConnectedChannel using existing ConnectedConnection.

pad_channel_id

Create a random number of dummy channel IDs so that the bootstrapped channel ID is random instead of being always channel-0.