Function ibc_integration_test::bootstrap::single::bootstrap_single_node

pub fn bootstrap_single_node(
    builder: &ChainBuilder, 
    prefix: &str, 
    config_modifier: impl FnOnce(&mut Value) -> Result<(), Error>
) -> Result<FullNode, Error>

Bootstrap a single full node with the provided ChainBuilder and a prefix for the chain ID.

The function would generate random postfix attached to the end of a chain ID. So for example having a prefix "alpha" may generate a chain with an ID like "ibc-alpha-f5a2a988"

The bootstrap function also tries to use as many random parameters when intitializing the chain, such as using random denomination and wallets. This is to help ensure that the test is written to only work with specific hardcoded parameters.

TODO: Due to the limitation of the gaiad command, currently parameters such as the stake denomination (stake) and the wallet address prefix (cosmos) cannot be overridden. It would be great to be able to randomize these parameters in the future as well.