Trait ibc_integration_test::chain::driver::tagged::TaggedChainDriverExt

pub trait TaggedChainDriverExt<Chain> {
    fn query_balance(
        &self, 
        wallet_id: &MonoTagged<Chain, &WalletAddress>, 
        denom: &MonoTagged<Chain, &Denom>
    ) -> Result<u64, Error>;

    fn assert_eventual_wallet_amount(
        &self, 
        user: &MonoTagged<Chain, &Wallet>, 
        target_amount: u64, 
        denom: &MonoTagged<Chain, &Denom>
    ) -> Result<(), Error>;

    fn transfer_token<Counterparty>(
        &self, 
        port_id: &TaggedPortIdRef<'_, Chain, Counterparty>, 
        channel_id: &TaggedChannelIdRef<'_, Chain, Counterparty>, 
        sender: &MonoTagged<Chain, &WalletAddress>, 
        recipient: &MonoTagged<Counterparty, &WalletAddress>, 
        amount: u64, 
        denom: &MonoTagged<Chain, &Denom>
    ) -> Result<(), Error>;

    fn local_transfer_token(
        &self, 
        sender: &MonoTagged<Chain, &WalletAddress>, 
        recipient: &MonoTagged<Chain, &WalletAddress>, 
        amount: u64, 
        denom: &MonoTagged<Chain, &Denom>
    ) -> Result<(), Error>;

    fn query_recipient_transactions(
        &self, 
        recipient_address: &MonoTagged<Chain, &WalletAddress>
    ) -> Result<Value, Error>;
}

A ChainDriver may be tagged with a Chain tag in the form MonoTagged<Chain, ChainDriver>.

It would implement the TaggedChainDriverExt trait to provide tagged version of the chain methods.

The tagged chain driver methods help ensure that the ChainDriver methods are used with the values associated to the correct chain.

Required methods

fn query_balance(
&self,
wallet_id: &MonoTagged<Chain, &WalletAddress>,
denom: &MonoTagged<Chain, &Denom>
) -> Result<u64, Error>

Tagged version of ChainDriver::query_balance.

Query for the balance of a wallet that belongs to Chain in the denomination that belongs to Chain.

fn assert_eventual_wallet_amount(
&self,
user: &MonoTagged<Chain, &Wallet>,
target_amount: u64,
denom: &MonoTagged<Chain, &Denom>
) -> Result<(), Error>

Tagged version of ChainDriver::assert_eventual_wallet_amount.

Assert that a wallet belongs to Chain would reach the target amount in the denomination that belongs to Chain.

fn transfer_token<Counterparty>(
&self,
port_id: &TaggedPortIdRef<'_, Chain, Counterparty>,
channel_id: &TaggedChannelIdRef<'_, Chain, Counterparty>,
sender: &MonoTagged<Chain, &WalletAddress>,
recipient: &MonoTagged<Counterparty, &WalletAddress>,
amount: u64,
denom: &MonoTagged<Chain, &Denom>
) -> Result<(), Error>

Tagged version of transfer_token. Submits an IBC token transfer transaction to Chain to any other Counterparty chain.

The following parameters are accepted:

A PortId on Chain that corresponds to a channel connected to Counterparty.

A ChannelId on Chain that corresponds to a channel connected to Counterparty.

The wallet address of the sender on Chain.

The wallet address of the recipient on Counterparty.

The transfer amount.

The denomination of the amount on Chain.

fn local_transfer_token(
&self,
sender: &MonoTagged<Chain, &WalletAddress>,
recipient: &MonoTagged<Chain, &WalletAddress>,
amount: u64,
denom: &MonoTagged<Chain, &Denom>
) -> Result<(), Error>

fn query_recipient_transactions(
&self,
recipient_address: &MonoTagged<Chain, &WalletAddress>
) -> Result<Value, Error>

Taggged version of query_recipient_transactions.

Query for the transactions related to a wallet on Chain receiving token transfer from others.