Module ibc_integration_test::framework::binary::chain

Constructs for running test cases with two chains, together with the relayer setup with chain handles and foreign clients.

Structs

RunBinaryChainTest

A wrapper type that lifts a test case that implements BinaryChainTest into a test case the implements BinaryNodeTest.

RunSelfConnectedBinaryChainTest

A wrapper type that lifts a test case that implements BinaryChainTest into a test case that implements BasicTest.

RunTwoWayBinaryChainTest

A wrapper type that lifts a test case that implements BinaryChainTest into a test case the implements BinaryChainTest.

Traits

BinaryChainTest

This trait is implemented for test cases that need to have two full nodes running together with the relayer setup with chain handles and foreign clients.

RelayerConfigOverride

An internal trait that can be implemented by test cases to override the relayer config before the relayer gets initialized.

SupervisorOverride

An internal trait that can be implemented by test cases to disable the spawning of the relayer supervisor.

Functions

run_binary_chain_test

Runs a test case that implements BinaryChainTest.

run_self_connected_binary_chain_test

Runs a test case that implements BinaryChainTest, with the test case being executed with a single chain that is connected to itself.

run_two_way_binary_chain_test

Runs a test case that implements BinaryChainTest, with the test case being executed twice, with the second time having the position of the two chains flipped.