Trait ibc_integration_test::framework::binary::channel::BinaryChannelTest

pub trait BinaryChannelTest {
    fn run<ChainA: ChainHandle, ChainB: ChainHandle>(
        &self, 
        config: &TestConfig, 
        relayer: RelayerDriver, 
        chains: ConnectedChains<ChainA, ChainB>, 
        channels: ConnectedChannel<ChainA, ChainB>
    ) -> Result<(), Error>;
}

This trait is implemented for test cases that need to have two full nodes running together with the relayer setup with chain handles and foreign clients, together with connected IBC channels with completed handshakes.

Required methods

fn run<ChainA: ChainHandle, ChainB: ChainHandle>(
&self,
config: &TestConfig,
relayer: RelayerDriver,
chains: ConnectedChains<ChainA, ChainB>,
channels: ConnectedChannel<ChainA, ChainB>
) -> Result<(), Error>

Test runner