Module ibc_integration_test::framework::nary::chain

Constructs for running test cases with more than two chains, together with the relayer setup with chain handles and foreign clients.

Structs

RunNaryChainTest

A wrapper type that lifts a test case that implements RunNaryChainTest into a test case the implements NaryNodeTest.

RunSelfConnectedNaryChainTest

A wrapper type that lifts a test case that implements RunNaryChainTest into a test case the implements NaryNodeTest<1>. i.e. only one underlying full node is spawned to emulate all chains.

Traits

NaryChainTest

This trait is implemented for test cases that need to have more than two chains running.

Functions

run_nary_chain_test

Runs a test case that implements NaryChainTest with a SIZE number of chains bootstrapped.

run_self_connected_nary_chain_test

Runs a test case that implements NaryChainTest, with one self-connected chain used to emulate many connnections.