Function ibc_integration_test::framework::nary::chain::run_self_connected_nary_chain_test

pub fn run_self_connected_nary_chain_test<Test, Overrides, const SIZE: usize>(
    test: &Test
) -> Result<(), Error> where
    Test: NaryChainTest<SIZE>,
    Test: HasOverrides<Overrides = Overrides>,
    Overrides: TestConfigOverride + NodeConfigOverride + RelayerConfigOverride, 

Runs a test case that implements NaryChainTest, with one self-connected chain used to emulate many connnections.

This works because IBC allows a chain to connect back to itself without the chain knowing it. Using this, we can emulate N-ary chain tests using only one chain and save the performance overhead of spawning many chains.

Note that with this, there is still performance overhead of establishing new connections and channels for each position, as otherwise the transferred IBC denoms will get mixed up. Some test cases also may not able to make use of self connected chains, e.g. if they need to start and stop individual chains.