Trait ibc_integration_test::framework::overrides::TestOverrides

pub trait TestOverrides {
    fn modify_test_config(&self, _config: &mut TestConfig) { ... }

    fn modify_node_config(&self, _config: &mut Value) -> Result<(), Error> { ... }

    fn modify_relayer_config(&self, _config: &mut Config) { ... }

    fn spawn_supervisor(
        &self, 
        relayer: &RelayerDriver
    ) -> Result<Option<SupervisorHandle>, Error> { ... }

    fn channel_port_a(&self) -> PortId { ... }

    fn channel_port_b(&self) -> PortId { ... }
}

This trait should be implemented for all test cases to allow overriding some parts of the behavior during the test setup.

Since all methods in this trait have default implementation, test cases that do not need any override can have an empty implementation body for this trait.

The trait provides generic implementation of the specialized traits such as RelayerConfigOverride. As a result, it is sufficient for test writers to only implement this trait instead of implementing the numerous override traits.

When a new override trait is defined, the same trait method should also be defined inside this trait with a default method body.

Provided methods

fn modify_test_config(&self, _config: &mut TestConfig)

fn modify_node_config(&self, _config: &mut Value) -> Result<(), Error>

Modify the full node config before the chain gets initialized.

The config is in the dynamic-typed toml::Value format, as we do not want to model the full format of the node config in Rust. Test authors can use the helper methods in chain::config to modify common config fields.

Implemented for NodeConfigOverride.

fn modify_relayer_config(&self, _config: &mut Config)

Modify the relayer config before initializing the relayer. Does no modification by default.

Implemented for RelayerConfigOverride.

fn spawn_supervisor(
&self,
relayer: &RelayerDriver
) -> Result<Option<SupervisorHandle>, Error>

Optionally spawns the relayer supervisor after the relayer chain handles and foreign clients are initialized. Default behavior is to spawn the supervisor using RelayerDriver::spawn_supervisor.

Test writers can disable the spawning of supervisor by overriding this method and making it do nothing and return None.

Implemented for SupervisorOverride.

fn channel_port_a(&self) -> PortId

Return the port ID used for creating the channel for the first chain. Returns the “transfer” port by default.

Implemented for PortsOverride.

fn channel_port_b(&self) -> PortId

Return the port ID used for creating the channel for the second chain. Returns the “transfer” port by default.

Implemented for PortsOverride.