Function ibc_integration_test::ibc::denom::derive_ibc_denom

pub fn derive_ibc_denom<ChainA, ChainB>(
    port_id: &TaggedPortIdRef<'_, ChainB, ChainA>, 
    channel_id: &TaggedChannelIdRef<'_, ChainB, ChainA>, 
    denom: &TaggedDenomRef<'_, ChainA>
) -> Result<TaggedDenom<ChainB>, Error>

A tagged version of derive_ibc_denom from the ibc module.

Derives the denom on ChainB based on a denom on ChainA that has been transferred to ChainB via IBC.

Accepts the following arguments:

A PortId on ChainB that corresponds to a channel connected to ChainA.

A ChannelId on ChainB that corresponds to a channel connected to ChainA.

The original denomination on ChainA.

Returns the derived denomination on ChainB.