Module ibc_integration_test::prelude

Re-export of common constructs that are used by test cases.

Re-exports

pub use crate::chain::driver::tagged::TaggedChainDriverExt;

pub use crate::chain::driver::ChainDriver;

pub use crate::error::handle_generic_error;

pub use crate::error::Error;

pub use crate::framework::base::HasOverrides;

pub use crate::framework::binary::chain::run_binary_chain_test;

pub use crate::framework::binary::chain::run_self_connected_binary_chain_test;

pub use crate::framework::binary::chain::run_two_way_binary_chain_test;

pub use crate::framework::binary::chain::BinaryChainTest;

pub use crate::framework::binary::chain::RunBinaryChainTest;

pub use crate::framework::binary::chain::RunSelfConnectedBinaryChainTest;

pub use crate::framework::binary::channel::run_binary_channel_test;

pub use crate::framework::binary::channel::run_two_way_binary_channel_test;

pub use crate::framework::binary::channel::BinaryChannelTest;

pub use crate::framework::binary::channel::RunBinaryChannelTest;

pub use crate::framework::binary::connection::run_binary_connection_test;

pub use crate::framework::binary::connection::run_two_way_binary_connection_test;

pub use crate::framework::binary::connection::BinaryConnectionTest;

pub use crate::framework::binary::connection::RunBinaryConnectionTest;

pub use crate::framework::binary::node::run_binary_node_test;

pub use crate::framework::binary::node::BinaryNodeTest;

pub use crate::framework::binary::node::RunBinaryNodeTest;

pub use crate::framework::nary::chain::run_nary_chain_test;

pub use crate::framework::nary::chain::run_self_connected_nary_chain_test;

pub use crate::framework::nary::chain::NaryChainTest;

pub use crate::framework::nary::chain::RunNaryChainTest;

pub use crate::framework::nary::chain::RunSelfConnectedNaryChainTest;

pub use crate::framework::nary::channel::run_binary_as_nary_channel_test;

pub use crate::framework::nary::channel::run_nary_channel_test;

pub use crate::framework::nary::channel::NaryChannelTest;

pub use crate::framework::nary::channel::RunBinaryAsNaryChannelTest;

pub use crate::framework::nary::channel::RunNaryChannelTest;

pub use crate::framework::nary::connection::run_nary_connection_test;

pub use crate::framework::nary::connection::NaryConnectionTest;

pub use crate::framework::nary::connection::RunNaryConnectionTest;

pub use crate::framework::nary::node::run_nary_node_test;

pub use crate::framework::nary::node::NaryNodeTest;

pub use crate::framework::nary::node::RunNaryNodeTest;

pub use crate::framework::overrides::TestOverrides;

pub use crate::relayer::channel::TaggedChannelEndExt;

pub use crate::relayer::connection::TaggedConnectionEndExt;

pub use crate::relayer::connection::TaggedConnectionExt;

pub use crate::relayer::driver::RelayerDriver;

pub use crate::relayer::foreign_client::TaggedForeignClientExt;

pub use crate::types::binary::chains::ConnectedChains;

pub use crate::types::binary::channel::ConnectedChannel;

pub use crate::types::binary::connection::ConnectedConnection;

pub use crate::types::config::TestConfig;

pub use crate::types::id::*;

pub use crate::types::single::node::FullNode;

pub use crate::types::single::node::TaggedFullNodeExt;

pub use crate::types::tagged::DualTagged;

pub use crate::types::tagged::MonoTagged;

pub use crate::types::wallet::TaggedTestWalletsExt;

pub use crate::types::wallet::TaggedWallet;

pub use crate::types::wallet::TestWallets;

pub use crate::types::wallet::Wallet;

pub use crate::types::wallet::WalletAddress;

pub use crate::types::wallet::WalletId;

pub use crate::util::assert::*;

pub use crate::util::retry::assert_eventually_succeed;

pub use crate::util::suspend::suspend;

Macros

debug

Constructs an event at the debug level.

error

Constructs an event at the error level.

eyre

Construct an ad-hoc error from a string.

info

Constructs an event at the info level.

warn

Constructs an event at the warn level.

Structs

ChainId

This type is subject to future changes.

ChannelId

ClientId

Config

ConnectionId

ForeignClient

PortId

SharedRegistry

SupervisorHandle

A wrapper around the SupervisorCmd sender so that we can send stop signal to the supervisor before stopping the chain drivers to prevent the supervisor from raising errors caused by closed connections.

Traits

ChainHandle

Type Definitions

SharedConfig