Function ibc_integration_test::relayer::transfer::tx_raw_ft_transfer

pub fn tx_raw_ft_transfer<SrcChain: ChainHandle, DstChain: ChainHandle>(
    src_handle: &SrcChain, 
    dst_handle: &DstChain, 
    channel: &ConnectedChannel<SrcChain, DstChain>, 
    recipient: &MonoTagged<DstChain, &WalletAddress>, 
    denom: &MonoTagged<SrcChain, &Denom>, 
    amount: u64, 
    timeout_height_offset: u64, 
    timeout_seconds: Duration, 
    number_messages: usize
) -> Result<Vec<IbcEvent>, Error>

Perform the same operation as hermes tx raw ft-transfer.

The function call skips the checks done in the CLI, as we already have the necessary information given to us by the test framework.

Note that we cannot change the sender’s wallet in this case, as the current send_tx implementation in CosmosSdkChain always use the signer wallet configured in the ChainConfig.

Currently the only way you can transfer using a different wallet is to create a brand new ChainHandle with the new wallet specified in the ChainConfig.

Alternatively, it is recommended that for simple case of IBC token transfer, test authors should instead use the transfer_token function provided by ChainDriver. That uses the gaiad tx ibc-transfer command to do the IBC transfer, which is much simpler as compared to the current way the relayer code is organized.