Type Definition ibc_integration_test::types::nary::chains::NthConnectedChains

pub type NthConnectedChains<const CHAIN_A: usize, const CHAIN_B: usize, Handle> = BinaryConnectedChains<NthChainHandle<CHAIN_A, Handle>, NthChainHandle<CHAIN_B, Handle>>;

A pair of binary ConnectedChains that are tagged by a Handle: CHainHandle and the const generics CHAIN_A: usize and CHAIN_B: usize.

Recall that binary ConnectedChains is tagged by two generic types ChainA: ChainHandle and ChainB: ChainHandle. For the case of N-ary chains, all elements must have the same type Handle: ChainHandle. But we want to still able to differentiate them when used as type parameter to ConnectedChains.

The solution is to tag each Handle with the const generic positions. So the first chain is MonoTagged<Size<CHAIN_A>, Handle>, which has a different type from the second chain MonoTagged<Size<CHAIN_B>, Handle>.

Since writing the fully qualified chain types are rather cumbersome, we use the type alias TaggedConnectedChains to refer to connected chains that are parameterized by const generics rather than the usual abstract type tags.